[Feature #111504652] (quickfix) Decouple serializers
