+++ codango/api/views/views.py
@@ -2,6 +2,13 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 
+# Import serializers
+from resources.serializers import ResourceSerializer
+from userprofile.serializers import UserProfileSerializer, FollowSerializer, \
+    LanguageSerializer, NotificationSerializer
+from votes.serializers import VoteSerializer
+from comments.serializers import CommentSerializer
+from pairprogram.serializers import SessionSerializer, ParticipantSerializer
 
 # Import models
 from resources.models import Resource
@@ -10,8 +17,8 @@ from votes.models import Vote
 from comments.models import Comment
 from pairprogram.models import Session, Participant
 
-# Import serializers
-from api import serializers
+
+
 
 # Default permission_classes = (permissions.IsAuthenticated,)
 
@@ -22,62 +29,61 @@ class ResourceViewSet(APIView):
     """
     def get(self, request):
         resources = Resource.objects.all()
-        serializer = serializers.ResourceSerializer(resources, many=True)
+        serializer = ResourceSerializer(resources, many=True)
         return Response(serializer.data)
 
 
 class UserProfileViewSet(APIView):
     def get(self, request):
         userprofiles = UserProfile.objects.all()
-        serializer = serializers.UserProfileSerializer(userprofiles, many=True)
+        serializer = UserProfileSerializer(userprofiles, many=True)
         return Response(serializer.data)
 
 
 class FollowViewSet(APIView):
     def get(self, request):
         follows = Follow.objects.all()
-        serializer = serializers.FollowSerializer(follows, many=True)
+        serializer = FollowSerializer(follows, many=True)
         return Response(serializer.data)
 
 
 class LanguageViewSet(APIView):
     def get(self, request):
         languages = Language.objects.all()
-        serializer = serializers.LanguageSerializer(languages, many=True)
+        serializer = LanguageSerializer(languages, many=True)
         return Response(serializer.data)
 
 
 class NotificationViewSet(APIView):
     def get(self, request):
         notifications = Notification.objects.all()
-        serializer = serializers.NotificationSerializer(notifications,
-                                                        many=True)
+        serializer = NotificationSerializer(notifications, many=True)
         return Response(serializer.data)
 
 
 class VoteViewSet(APIView):
     def get(self, request):
         votes = Vote.objects.all()
-        serializer = serializers.VoteSerializer(votes, many=True)
+        serializer = VoteSerializer(votes, many=True)
         return Response(serializer.data)
 
 
 class CommentViewSet(APIView):
     def get(self, request):
         comments = Comment.objects.all()
-        serializer = serializers.CommentSerializer(comments, many=True)
+        serializer = CommentSerializer(comments, many=True)
         return Response(serializer.data)
 
 
 class SessionViewSet(APIView):
     def get(self, request):
         sessions = Session.objects.all()
-        serializer = serializers.SessionSerializer(sessions, many=True)
+        serializer = SessionSerializer(sessions, many=True)
         return Response(serializer.data)
 
 
 class ParticipantViewSet(APIView):
     def get(self, request):
         participants = Participant.objects.all()
-        serializer = serializers.ParticipantSerializer(participants, many=True)
+        serializer = ParticipantSerializer(participants, many=True)
         return Response(serializer.data)
