+++ codango/comments/serializers.py
@@ -1,0 +1,12 @@
+from rest_framework import serializers
+
+from models import Comment
+
+
+class CommentSerializer(serializers.ModelSerializer):
+    """Comment Serializer"""
+
+    class Meta:
+        model = Comment
+        fields = ('author', 'resource', 'content', 'date_created',
+                  'date_modified')
